Avoid copying the .o files to ghci when .a is not a thin archive
Summary:
We need copy the .o files to ghci when .a is a thin archive as it references
them by relative path. But we .a file is a normal archive, we should not do so
to avoid wasting space.

The previous version introduced a bug when thin archive is used, as we
removed the copied objects when we were later copying the interfaces. This
version fixed it.

Reviewed By: karen

shipit-source-id: d9dcdace1d8425051fc2d96571f063e75aa75cdc