Ensure that isImmutable is properly implemented for BuiltInProviderInfo
Summary:
Previously we just always said that BuiltInProvider is mutable. This is inaccurate, and can be generalized. If all of the values in skylark fields are immutable, this provider is considered "Immutable". We also add in a fast path on this check because once an object is immutable, it cannot be made mutable again.

This is necessary to implement an invariant where all objects that come out of the RA framework's implementation functions are immutable.

Reviewed By: cjhopman

shipit-source-id: f4b7dfc18d6801b4c8d123109d4df77b1f25e6b6