Allow mutable objects to be passed to user defined provider info
Summary:
Previously passing a mutable value into a provider would fail with an error. This caused problems when passing a struct in, as that was considered mutable, and there was no user facing way to make it immutable. It also led to us copying data more than necessary (see the now removed asDeepImmutable call). Instead, rely on mutability again to ensure that this object freezes properly at the end of impl execution.

Note: This does mean if a list is passed to a provider, that list can still be mutated. We may, in the future, want to have a keyword to provider() to change this behavior.

e.g.
```
FooInfo = provider(fields=['bar'])
some_list = [1,2,3]
f = FooInfo(bar=some_list)
some_list.append(4)
# At this point f.some_list == [1,2,3,4]
# Previously it would have been [1,2,3]
````

Reviewed By: cjhopman

shipit-source-id: 79b0af464a96cf858d658dc0d6df7c63a889d72b