always allow 'untracked headers' for infer capture
Summary:
We have cxx.untracked_headers=ignore on various places for Infer cxx capture.
They never work reliably for the various mix of mode files we include.
So, for CxxInferCapture, hardcode the use of HeaderVerification.Mode.IGNORE, since this is the only setting we are interested in.

This fixes compilation errors on fbandroid/cpp

Reviewed By: jberdine

shipit-source-id: f6ea2102a844c10ba93780d9104a14aabe46fdff