UnconfiguredTargetNode.compatibleWith field
Summary:
Store `compatible_with` explicitly in a field of `UnconfiguredTargetNode`.

This is done to avoid interference of manual handling of `compatible_with` with two-phase coercion implementation.

Also it feels right to store more information in typed fields instead of map by string even if `compatible_with` exists only for build rules, and not configuration rules.

Reviewed By: bobyangyf

shipit-source-id: c8c7c4eee50e29f7acc2e152b894d51208661c03