Skip unnecessary creation of String objects
Summary:
We have a very hot loop that iterates over the SYMTAB which can have several millions of entries (8-20 million for a medium to large sized binary). Perf measurements reveal significant overhead in the creation of `String` objects (decoding + allocations) which are completely unnecessary.

In this diff, we remove the usage of `String` objects for non-OSO SYMTAB entries. Usages of `String` for OSO entries will be removed in the next diff.

Reviewed By: stepancheg

shipit-source-id: 34007e2620034249c9de84f2a7bfaadb026e1d13