Remove incorrect write to ARTIFACT_METADATA
Summary:
On D18332196 we changed Buck to use a single ARTIFACT_METADATA file (in contrast with having one file for OUTPUT_SIZE, one for OUTPUT_HASH, etc).

We are incorrectly writing the OUTPUT_SIZE directly to the new ARTIFACT_METADATA file. This seems to be left by mistake.

It could be the root cause of P126827042, as seen on https://our.intern.facebook.com/intern/buck/build/6b37db54-63f0-45b7-94ec-82e16e32072c

This is an attempt to help S196742

Reviewed By: IanChilds

shipit-source-id: a1dce25f27c89f0276f415919bd59481665c51bf