Unused dep finder should also use exported deps for first-order-deps check
Summary:
We already say that if a dep A is only "used" in the graph because it exports some used dep B, but B is also specified as a first-order-dep, then we will mark A as unused (as we can safely remove it).

This diff makes it so that we have the same behaviour if B is specified as a first-order exported dep, rather than dep.

Reviewed By: brianduff

shipit-source-id: 98778d165697ee94acf8894eb444836da80fdf27