Remove umbrella header generation
Summary:
Buck will currently generate an umbrella header if one is not included in the module. Since we are moving away from using umbrella headers entirely, we no longer need this feature.

This is the first step towards removing the `umbrella_header` modulemap mode entirely, and using the `headers` mode exclusively.

Reviewed By: milend

shipit-source-id: 673ae914cb749db0280260f4a1415b7d822c6c14