modular does not imply CLANG_ENABLE_MODULES for non-framework targets
Summary:
A library can //define// a module for dependents to consume without being compiled with modules enabled, but these two behaviors are currently coupled in `buck project`. This is not the case in `buck build`, so if verification is only done with the latter, this can break the build for module compatibility issues when `buck project` is used instead.

There's a comment that implies that we need it for framework modules, so keep it there.

Reviewed By: milend

shipit-source-id: cff2cf498199188841f9eb6dd24bce7da8cd732a