add hint to mark that inputs must be regular files
Summary:
I noticed user can use directories in `raw_headers` or `srcs` of cxx rules, for example.

In case of `raw_headers` in works in bit unexpected way and causes some other tools to fail. But in case of `srcs` it fails during the compilation while this will fail much earlier during target graph construction (e.g. it will fail on `buck query` or `buck targets`).

It will be used in the next diffs.

Reviewed By: stepancheg

shipit-source-id: 685f3ee2ab7abb733d1a70a542eefecc2864c48c