Significantly speed up initial buck exo install
Summary:
The files copied to /data/local/tmp/exopackage during buck install are chmoded to make them visible to the installed app. The chmod was being done via adb, which turned out to be excessively slow when there are a lot of files.

On modern devices & emulators, it seems like this chmodding is no longer necessary (on both a modern physical device I tried, and an emulator running N, the exopackage files were already created readable by other users). However, in case there are devices where this is still required, I preserved the chmoding, but moved it to run inside the agent. It's fast enough here to take a negligible amount of time (I timed it with and without the chmods).

Reviewed By: IanChilds

shipit-source-id: a53e79ee9e33c9f60c1fe1d31d90be0b59f60b30