Write DirArtifactCache metadata directly
Summary:
This is similar to D20050146 - we see wins from not writing to one file and then moving it into place. (https://fb.workplace.com/groups/1037883282914066/permalink/2942413692461006/).

I believe that this is safe, because if the operation does get interrupted then the metadata will not be well-formed, and when we go to read the metadata (https://fburl.com/diffusion/0rtigjc0), we'll get an IOException and that will go down as a cache error, which is correct.

Reviewed By: asp2insp

shipit-source-id: 2ebb124c638d8e01700e33ef4712c31536716923