Add default_outs to genrule
Summary:
Since we're not going to remove "single output list restriction", genrule's default outputs can't return the set of all named outputs. So we need to figure out new default outputs for genrule

This diff adds `default_outs` to genrule, which allows users to specify default outputs

`default_outs` should always be present if `outs` is present. However, since the repo usages right now don't have `default_outs`, I haven't added this check yet. The behavior in this diff is that default outputs are empty if `default_outs` is not specified, otherwise the default outputs would be `default_outs`. After submitting this diff and after this diff gets into the version bump, I plan to change the repo usages to use `default_outs`, then we can add a `Preconditions` check for asserting that `default_outs` must be present if `outs` is present and remove the logic for empty defaults.

Will fix more tests with `default_outs` in follow-up diffs.

Reviewed By: cjhopman

shipit-source-id: d8dd04f660c6c8845f00f545d9f220833e6a664f