Add chmoding to the native agent
Summary:
The native agent is used on Lollipop and below devices. Despite documentation to the contrary, it also doesn't chmod files, but instead was relying on buck doing this through adb shell commands. When the optimization in D20242616 is turned on, this no longer happens, leading to crashes.

This diff fixes the crashes by adding chmod() calls into the native agent.

Reviewed By: IanChilds

shipit-source-id: 6cd75318a5df9d30f80daf6a171017ad9f1f9e4c