Remove benchmark binary target
Summary:
Java binaries are large and take long to create (extra couple seconds at local build).

And this particular binary is never used to my knowledge.

Also there is a dozen of other benchmark which are not available as binaries.

(Also we Caliper benchmark is not as good as JMH benchmark we also use.)

Reviewed By: KapJI

shipit-source-id: e3c8b770a1232fe618909918f757b239f6ad5827