Pass underlying module via -Xcc -I
Summary:
This diff fixes bidirectional Swift/Objective-C interop in modular libraries. Currently, the build will fail when attempting to import the `-Swift.h` header, because the regular modulemap is used when building in Xcode, despite us attempting to specify `objc.modulemap` with a VFS overlay.

It appears that `-ivfsoverlay` does not work when Objective-C code is included via `SWIFT_INCLUDE_PATHS`, which translates to Swift's `-I`, not `-Xcc -I`. The former is necessary to import Swift code, but since we're importing the module's Objective-C code, that isn't an issue. There is also a (newer?) `-vfsoverlay` option that does not need `-Xcc`, but it has the same behavior.

`ImmutableList.of` also seems shorter and more readable, any reason not to use it if there's no branching builder logic?

Reviewed By: milend

shipit-source-id: 2a67522d8fb8d411c5e3925d1ef1c8b5f9f7d459