Remove -Swift.h from headers modulemaps in project
Summary:
The `headers` modulemaps generated by `buck project` currently include the `-Swift.h` header, like this:

```
module Mixed {
    header "Hello.h"
    header "Mixed-Swift.h"
    export *
}
```

This breaks compiling Swift code in a mixed module, since the header does not exist yet. So, we need to remove it from the set of headers to add when generating the modulemap.

This doesn't seem like the best way to do this, and I'm going to see if there's a better way, but I'm putting this up for early review for now to get feedback, especially on the next diff.

Reviewed By: milend

shipit-source-id: fd9cc6c6ad9c74a8ae0e208a9824f4937dc87341