Update visibility checking to be specific on the violation
Summary:
Previously, regardless of whether it was visibility or within view that was violated, we would report on the error a target not being visible.

This updates the behavior to report specifically on whether the violation is "not visible" or "not within view".

Additionally API changes also set up for performing an exhaustive visibility check.

Reviewed By: stepancheg

shipit-source-id: 0f6447d3ef58ec4dabcf40b100768cdac6963688