Introduce `buck cquery`
Summary: Current `buck query` technically runs on the configured target graph but it was created in an era where there was no distinction between the unconfigured and configured target graph, so some of it's semantics haven't aged well. This new command is meant to give us a clean slate, where we can mess with the output indiscriminently and not worry about all the legacy uses we'll be disrupting. Then we can start migrating consumers over to the new query.

Reviewed By: cjhopman

shipit-source-id: 478ad95b5785ae744cf2642d8b93959078b96871