Make UDR work with 'versions'
Summary:
When versions are enabled, we take a different path to create ConstructorArg objects than we do with versions disabled (see TargetNodeTranslator.java). In order to actually create a new instance of a SkylarkDescriptionArg, we need to know not just its class, but also which UDR is being referred to. This information is not directly available from the type coercer, however in TargetNodeTranslator, we have the original ConstructorArg instance, which is enough to make a new DTO builder.

This change does a few things:
- Renames 'getConstructorArgDescriptor' to 'getNativeConstructorArgDescriptor' to make it clear that a Class alone cannot get you a DTODescriptor for UDR
- Adds an interface to SkylarkDescriptionArg that allows us to create a DTODescriptor based on the old arg.
- Centralizes the DTODescriptor/RuleDescriptor logic into SkylarkDescriptionArg, and updates KnownUserDefinedRules to use it
- Updates the type coercer to allow fetching ParamInfo mapping directly. There are a number of places that were fetching a whole DTODescriptor just to fetch ParamInfos. We can keep doing that for most native rules, but for UDR, we already have this info available directly on the existing SkylarkDescriptionArgs. This allows us to 1) not create unneeded intermediate objects, and 2) not need to use what is now getNativeConstructorArgDescriptor

Reviewed By: bobyangyf

shipit-source-id: 81a7aba7ffbab776971390c8c29d49cec5373bfd