Explicitly specify annotation processors to KAPT - with config option
Summary: Try this again. When I landed this before it broke the version bump, but now it has a config option which means that the default behavior is what we have right now. When we try to change the config option, then we can see fbsource failures and fix them.

Reviewed By: cute-jumper

shipit-source-id: 9d6259cc48a9ee56056b1fbbca68c202d60f8468