fbsource: remove illegal .watchmanconfig files
Summary:
We only allow the singular `.watchmanconfig` file at the root of the
repo; it is too expensive to allow for overlapping watches at arbitrary repo
locations (the overhead is exponential rather than linear in the number of
overlaps), and with Eden, it is simply not supported.

These `.watchmanconfig` files all appear to be working around this FB specific
requirement by creating invalid empty configuration files.

The correct solution in opensource projects is to not use a `.watchmanconfig` at all,
as the default behavior outside of FB is to allow the `.git` directory to act as the
project root.

In addition, correct usage of `watch-project` instead of the deprecated `watch` allows
for watching subdirectories of fbsource.

For projects that need to run tests inside the FB environment with an isolated
instance, the recommendation is to have your test set up the test location in a
temporary directory fixture outside the repo and set up the `.watchmanconfig`
file on the fly as part of that setup.

Reviewed By: simpkins

shipit-source-id: 3c00338d4afd3ea01fb63eda1bcc590d688f99ec