cancel action if it fails and propagate it in appropriate future
Summary:
There was a bug - while executing RE action, grpc stack could have crashed but wrong future was set with failed result (setException) and so it was never propagated up to the awaiting function and build were getting stuck.
This diff just changes which future buck sets to failed, now it uses the correct one.

Few words on how it looked for users and in prod:
Build hangs only when there i something wrong with connection with Engine, like connection died or there were significant packet loss so netty decided to re-establish connection. This doesn't happen too often for short builds (less than 30min) because network is reliable, Engine is waiting for connections to be closed and so everything works pretty smoothly.
The bug hits us in cases:
* super long build/connections
* flakey network or any other issue with network
* **engine crashed**

and so on.

The extreme test case for this diff is to run short build in remote mode with no engine running. Buck gets stuck forever, even though grpc/netty gives up pretty soon. This diff changes this behaviour to:
* **catch failures on network and retry them on netty level**
* **catch fatal failures on network and proparage failure to the upper stack** - crash action immediately when grpc gives up

No more hanging builds in remote and hybrid mode!

Steps of the investigation are described here: https://fb.workplace.com/groups/1208881079252335/permalink/1640040789469693/

Reviewed By: pranjalssh

shipit-source-id: bffb5f7ee50eb0e8e0a05c51d4fc43793723a3ab