BuildCommandForProjectGenerators
Summary:
When IDE generator invokes `build`, targets must be built with specified configurations.

This diff introduces `BuildCommandForProjectGenerators` which accepts a list of `BuildTarget` instead of a list of targets as string. (Existing `BuildCommand` is still a handler or `buck build` command invocation.)

This is a preparation to further refactorings, several changes will be needed.

This diff would be useful to land earlier to avoid merge conflicts later.

Reviewed By: bobyangyf

shipit-source-id: 711c03fc97be3121b815404dd59b0408f3e34654