Use symlink trees for modular libraries in project
Summary:
We're currently diverging from the behavior of `buck build` here, and it causes build failures when using `buck project` instead. In `build`, when we a library depends on a modular library, it uses a symlink tree - specifically, the "symlink tree with modulemap" flavor. However, in `project`, we always use headermaps. If we actually enable modules in Objective-C, the header-map-based imports do not work. To resolve this, we need to add the symlink trees to the header search paths for Objective-C code.

We already had a function to do this for Swift, so that was reused. This is currently gated behind a new `--config` flag.

Reviewed By: milend

shipit-source-id: 499e1270475634afe5b71c38f2709fa01a900a2d