retry UNKNOWN in grpc
Summary:
We have seen UNKNOWN errors mostly on events like "Channel closed",
which, usually, should be treated as UNAVAILABLE. Apparently, in some case, if
connection issue happens in the middle of parsing/reading request, the request
might end up with UNKNOWN status.
RE, on the other hand, doesn't care much if request was executed twice even if the
first one was successful.

Reviewed By: pranjalssh

shipit-source-id: 11de736eef9e892b4ae7303c528803090f892f35