Validate existence of recorded cache artifacts
Summary:
In this diff, we check whether the cache artifact metadata matches up with the filesystem. More specifically, we check that any recorded paths and hashed paths exist on the filesystem.

The hashed paths will be a larger set than the recorded paths, as it includes all files (e.g., a recorded path might be a directory).

Reviewed By: cjhopman

shipit-source-id: c691216a4696591fa91e6e55ad6fed04c99c4565