Correctly invalidate DirectoryListCache when new folders are created
Summary:
Whenever a new folder is created, Watchman informs us by sending a WatchmanEvent with a kind of CREATE, with paths to files in the created directory. Previously, we were only invalidating the directory list cache entry for the containing folder of that file upon receiving that event. That alone is not sufficient, since the parent folder of that file may itself be a new folder that was previously unknown to the DirectoryListCache.

To remedy this, this commit walks the path upwards towards the root, invalidating the cache of all entries that did not previously know about the directory represneted in the new path. This ensures that we invalidate the cache for the immediate parent of all directories that were created prior to the insertion of the entry into the cache.

Reviewed By: irenewchen

shipit-source-id: a2a357b9c0b9c2f12e0dcdb8be3cabdcdb58c203