Parsing only failures in the gtest report
Summary:
CxxGtestTest is assuming that the children of a test case only contain a failure node. This prevents any other data to be included in the gtest report.

This diff addresses the issue by parsing only failure child node. It will makes the assumption (as previously that there is only 1 failure child node, which probably holds for now).

Reviewed By: bobyangyf

shipit-source-id: f1800930b896626a1b96339e3d90e935b0f47e69