Move anything `QueryTarget` related from `CommandHelper` to `QueryCommand`
Summary: The `QueryTarget` class should be considered an implementation detail of the query classes, not something generic. Not only that, but the way which `CommandHelper` prints these targets is incompatible with `ConfiguredQueryCommand`, meaning that it's really only useful for `QueryCommand`. In that case, move it to the class which actually uses it.

Reviewed By: cjhopman

shipit-source-id: 8e5f0213652c4e86a55746dba452e7c9b26e9108