Move the majority of logic from AbstractQueryCommand to QueryCommand
Summary:
This may deserve some background.

The first version of this diff (which is now doomed to sit on my laptop forever) simply moved everything related to `MergedTargetNode` from `AbstractQueryCommand` to `QueryCommand`. The thinking was that `MergedTargetNode` was the class responsible for the problems we are trying to solve with `cquery` - therefore it could be a signpost for code which doesn't belong in `ConfiguredQueryCommand`.

The result of that effort was to make most of the output formats into abstract methods - `printDotOutput`, `printThriftOutput`, `printRankOutput`, etc. The only two left in `AbstractQueryCommand` were `printListOutput` and `printJsonOutput`. I was all set and ready to submit the change when I realized that `CommandHelper` was taking `QueryTarget`s. Weird, I thought, `QueryTarget` should be an implementation detail of query. I then found out that `CommandHelper` was _also_ generating output we don't want in `ConfiguredQuery` - it was taking the list of nodes returned from the query and then making a set keyed off of `BuildTarget.getFullyQualifiedName()`[1]. This meant that everything which used `CommandHelper` had to get out of `AbstractQueryCommand`, which meant the last two `printXXXX` methods had to go as well.

So this diff moves everything printing related. We introduce two high level abstract methods as entry points into printing logic, and then move all the current implementations into `QueryCommand`. This leaves `AbstractQueryCommand` with very little logic, just some common CLI args and the logic around set syntax (`%s` and `%S`). Given that the actual querying implementation lives inside `BuckQueryEnvironment`, that sounds about right.

[1] The fact that the `CommandHelper` methods are `QueryCommand`-only also makes me think they should be on `QueryCommand` itself. I will do that in a later diff.

Reviewed By: bobyangyf

shipit-source-id: 1b90e1750779eac04e171c5b77b7308184b5f490