Remove defunct CoreSimulator service path check from InstallCommand
Summary:
This dates to the most ancient versions of iOS support for `InstallCommand`. It attempts to invoke `launchctl print` to fetch data about the currently running CoreSimulatorService so it can restart it if its path does not match the Xcode developer directory path.

Sample out put for that `launchctl print` command is here: P127866784

However, in the ensuing five years this has broken for two reasons.

1. The path is now always in `/Library/Developer/PrivateFrameworks/CoreSimulator.framework`. Thus Buck incorrectly assumes it *always* has to restart CoreSimulatorService since that path never matches the Xcode developer directory.
2. Parsing the output of `launchctl print` fails because it now contains control characters of some sort. (Maybe color codes? I'm not sure.) To be fair, there is a stern warning in the `man launchctl` manpage: `IMPORTANT: This output is NOT API in any sense at all. Do NOT rely on the structure or information emitted for ANY reason.`

Remove it entirely. I think we can safely assume that if there's a running simulator service we can use it.

Reviewed By: williamtwilson

shipit-source-id: b5b98948462769075f6b43b1be9709a6483a2363