Fix NPE in TargetGraph.getSubgraph
Summary:
getSubgraph throws an NPE when the TargetGraph doesn't contain the unflavored version of a node, just the flavored version.
This change guards against this NPE by checking for contains before adding to the map.

I'm not 100% sure this is the right solution since D2552014 seemed to try to ensure the unflavored node is always present,
but that was a long time ago and there's been significant changes wrt configured graphs and versioned graphs etc since then.

Reviewed By: andrewjcg

shipit-source-id: 949e8aa120b25b642e88461e03f0ca4b4b8dded5