Relax typing of `BuckQueryEnvironment` and friends
Summary:
The typing of these functions is really wacky.

In many places, we have generics like `T extends QueryTarget`, and then functions and such will use `<QueryBuildTarget>`. This is a lie, because even if functions say they only take `QueryBuildTarget` as input the actual `BuckQueryEnvironment` doesn't know that and happily passes it a `QueryFileTarget`, because according to the environment the functions take `QueryTarget`. This leads us to situations where correct looking code is causing bugs because it's getting types that it's not expecting. You can easily create queries that cause Buck to crash with a stack trace and that is generally due to some function signature saying it takes `QueryBuildTarget` but the environment passes it a file.

This was a rabbit hole I jumped down due to the incorrect typing on `OwnerFunction`, which says it takes a `QueryBuildTarget` as input when in reality it's the only function that takes a `QueryFileTarget`.

Anyway, we can fix this by not lying to Java about what types are being passed to us. We only take `QueryTarget` - period.

Reviewed By: stepancheg

shipit-source-id: 4435596874bd67cd7832f607a512a38f3f11743b