Use `BuckQueryEnvironment` for cquery
Summary:
This class does most of the heavy lifting for the query, so we are able to reuse all that logic.

The implementation of `runWithoutHelp` and most of the printing logic is straight out of `QueryCommand`, with the caveat that it uses a different method to generate output. While it's a bit of copy pasta now I don't expect it to stay this way in the future, once we start getting ito the nitty gritty of cquery. For now, I just want to make an integration test pass. Any test. Please.

Reviewed By: stepancheg

shipit-source-id: 4ce2b759f0b05d751b057ca62283beda3354a9ad