Test that target_sdk_version works for Objective-C in binaries
Summary: The file `Test.m` is used to verify that specifying a `target_sdk_version` for a library adds the correct flags, by using a function gated to a newer target version. However, we're not testing it in binaries, only the Swift version. Likely missed due to already having `main.m`.

Reviewed By: milend

shipit-source-id: e540db84441e35abcdc045946bb453537a4348ae