compatibility with future infer CLI
Summary:
D20362641 deleted `--issues-txt` so the "pretty print" option will now crash. D20623613 is about to also delete the `--report-hook` option. Switch the behaviour of pretty printing to enable/disable `--quiet` instead, which is a more stable API.

D20623613 also renames bugs.txt to report.txt so we agree which one was better.

Reviewed By: artempyanykh

shipit-source-id: fab987507a5594adabaa3d5d9214419beb1ae3ec