Match "go test" behavior in more cases
Summary:
When running test binaries under "go test" the go tool ensures the
working directory of the test process is set to the package directory. The tool
does this to make sure that test authors can access test data in the package
directory. Buck instead ensures that any resources defined on the test target
are available in the same directory as the binary. Changing the working
directory of the test to the directory of the binary will have the same effect.

Before this change, buck's handling of resources defined on tests differed in
case of using the internal test runner or an external test runner. This diff
consolidates the behavior. Depending on the final output of the go binary for
the test binary only should make running these tests on remote execution
simpler in the future as well.

Note: buck provides an option to use `testmaingen` from the platform as defined
in the configuration. Those test targets will not have the "change dir" fix.
The internal runner will continue to set the working directory for backwards
compatibility.

Reviewed By: andrewjcg

shipit-source-id: 50ddcb781f07cc1cc246254a000e1bbf79c4c832