Relativize METADATA_PATH to the repository root
Summary:
The Buck client reads and writes a file METADATA_PATH which is used to communicate metadata on the execution of rules that ran remotely. Buck expects this file to be at the root of the cell that is currently being built. METADATA_PATH is itself a relative path, but it is relativized to the cell root and not the repo root.
This commit explciitly relativizes METADATA_PATH to the repository root by relativizing the absolute path with the cell path prefix.

Reviewed By: pranjalssh

shipit-source-id: 52b4b528dc7781c11fc45e39cb89f17d6c1b8ba8