Fix potential rulekey collisions with Aapt resources
Summary:
`aapt2_link` step is failing intermittently for release builds because of incorrectly cached `aapt2_compile` artifacts.

The incorrect artifacts appear to be `aapt2_compile` rules from the same or very similar build, but with files from a different index directory.

We are caching different `aapt2_compile` artifacts with the same rulekeys but different output hashes.

The most likely scenario is that there are multiple builds with identical resources but a different ordering for `resDirectories`. This would cause `resource_filter` rulekeys to not change when the output of the rule changes, and cause `aapt2_compile` rules with the same rulekeys to contain different resources.

If the above scenario is correct, hashing an ordered list of path names would be enough to fix it, rather than SourcePaths, which will hash all of the contents as well and which might be slow. Since this only affects release builds, we should err on the side of correctness.

Reviewed By: IanChilds

shipit-source-id: b0b15866f793b99dfb05744f0c8394498341bf5d