Integrate LogStreamFactory with Buck's LogFileHandler Attempt #2
Summary:
Integrate LogStreamFactory with Buck's LogFileHandler.

The break is caused by the returned object in FileOutputStreamFactory. There is a subtle difference between `return new FileOutputStream(...)` and `return Files.newOutputStream(...)`.

The latter upon closing the stream will close the ByteChannel to which it writes to, which doesn't play so well with ReferenceCountedWriter i.e. closing a ReferenceCountedWriter a second time will attempt to flush and close this ByteChannel again, which results in exception to close writer.

Reviewed By: mykola-semko

shipit-source-id: a018f8796d8742f8824d1e8389641c49867f3a8f