Update invalidatePath API to take whether to invalidate manifests
Summary:
Followup from [D18647934](https://our.intern.facebook.com/intern/diff/D18647934/?transaction_id=738366273339582&dest_number=94880584), to make cache invalidation more performant.

When invalidating a PACKAGE file, we want to invalidate its parsed file manifest but not the parsed file manifest of child PACKAGE files that depend on it. To support this we need to provide to any recursive calls on dependent files on whether to invalidate the manifests.

This diffs updates the API and the logical changes occur in D19268230

Reviewed By: bobyangyf

shipit-source-id: c4bbed5ce4017087e45158ba3382bc6c376e229d