Fix misbehavior with --version, -V
Summary:
This code looks for "--version" or "-V" in sys.argv. We have some cases where a user does `buck run foo:bar -- --version` from a subdirectory. This causes the root to be incorrectly detected as the CWD, rather than the real root.

Remove this special logic. It was likely written as a speedup, however the cost of a single stat in the common case (where a user is at the root), should be easily absorbed, as we read .buckconfig in a number of other locations.

Reviewed By: mykola-semko

shipit-source-id: 682b3149d0457cc6e5746cb33b49dc4ce3473f40