Look at exported_provided_deps when doing unused deps checking
Summary:
This fixes an existing bug - the behavior for `exported_provided_deps` is now the same as for `exported_deps`, which is what we want.

Suppose we have a target A, which depends upon B. B has an exported provided dep C. If A needs C in order to compile, but does not depend upon it directly, then before this diff we will mark B as unused, which is incorrect as removing it will cause A to fail to compile. After this diff, B will be correctly identified as "used".

shipit-source-id: cf6f9ac95dfc50ddaa7e2db88be6ef90850cddbe