Update package invalidation logic
Summary:
Followup from [D18647934](https://our.intern.facebook.com/intern/diff/D18647934/?transaction_id=738366273339582&dest_number=94880584), to make invalidation more accurate.

Previously, we would set a build file to be dependent on the package file in its directory and all parent directories. While this is effectively true (through optional package inheritance), it is not exactly true.

Build files are dependent on the package file in its directory. A package file is dependent on its parent package file. Recursively, the parent package file is dependent on its own parent package file.

Initially, we believed this caused n^2 operations when invalidating a package file. The hypothesis that in directory structure A/B/C, invalidation of A/PACKAGE would cause targets in A, B and C to be invalidated, as well as invalidate B/PACKAGE which would invalidate targets in B and C, etc. However, this is not true, as we never set B/PACKAGE to depend on A/PACKAGE.

This diff establishes this correct relationship, and updates the invalidation to be performant. This is achieved via:
1. Setting a build file to depend on the package file in it's directory (removing dependencies on all parent package files)
2. Setting a package file to depend on it's parent package file (previously missing)
3. Storing empty package files (previous diff D19267607) for parents package files that don't exist (such that relationships between parent/child are established)
4. Updating invalidation such that when package files are invalidated, nodes of targets defined in child build files are invalidated without invalidating the child package manifest themselves.

Reviewed By: bobyangyf

shipit-source-id: 5965ad0cbbddafc6a01264c800a70855ab5f3ce7