Remove unused parameter
Summary:
We want to do validation at two specific places: upload and download. At the download callsite, we have access to the extract files list but that list is actually an unused parameter.

Rather than introducing another `validateArtifactForUpload()` method that takes no parameters, we're simplying the code and removing the unused parameter. If it needs to be used, it can be done in the future.

Reviewed By: mykola-semko

shipit-source-id: 62aca8d7625ed649baec1bfccec2ad8e83a770bb