Add artifact validation to BuildCacheArtifactUploader
Summary: Add validation code when we upload cache artifacts. There's a potential for corruption between the metadata and what gets uploaded. That's because the files to be uploaded get determined by traversing the recorded output paths and if those are directories, another process or Buck could have changed the contents of any directories which were recorded as output paths. The metadata that would have been recorded would include hashes of all files, so a discrepancy can occur there.

Reviewed By: cjhopman

shipit-source-id: 2b189a1fe55ecab1944fe8694f2a5696b28e243d