Introduce uploader interface to be used by BuildCacheArtifactUploader
Summary:
`BuildCacheArtifactUploader` hardcodes the usage of `ArtifactUploader` which actually tries to hit the network, use the event bus and so on. This makes it much harder to write a test for `BuildCacheArtifactUploader`.

Here, we introduce a `NetworkUploader` interface which `BuildCacheArtifactUploader` uses to perform the actual upload. In the next diff, we will mock `NetworkUploader` in test conditions.

Reviewed By: cjhopman

shipit-source-id: f1b266a203b90afc289fa428d1a2d199f6dd323f