Implement #infer-java-capture flavor for InferJava build rule
Summary:
The flavor allows to run `infer capture` via `buck build //[library-name]#infer-java-capture`.
The output of the build rule is exactly `infer-out`.

When used with proper xplat-dist of infer, the rule can be execution remotely.

Note: the output is *exactly* infer-out, which means that some cruft like `tmp/` dir is also in the output (while ideally it should be just under scratch dir in `buck-out/bin`). We can improve this later on.

Reviewed By: ngorogiannis

shipit-source-id: d0427ecb6e07454bcd3ec0ce253e342b08362b54