Fix cquery test data
Summary:
The current testdata setup is incorrect - `//bin:mac-bin` transitively depends on `//lib:bar` which has a `select` for it's sources that has no `mac` case. This causes a parse error if you try to configure the target for the mac configuration.

This diff fixes the problem by adding a DEFAULT case

Reviewed By: stepancheg

shipit-source-id: 1859233800f3f57d32090294c569dcadf1e046d7