Do not override toString in SkylarkUserDefinedRule
Summary:
Override was done to remove star from `toString` output, but star is perfectly correct way to print function signature when all parameters are named.

For example, in Python 3:

```
$ python3
>>> def f(*, a, b): pass
...
>>> help(f)
f(*, a, b)
```

But the change is done to upgrade Starlark, patched code depends on Starlark internals which are not available in newer Starlark.

Reviewed By: philipjameson

shipit-source-id: 7b6b6e44458eb60ef937d37f30fb63e3e1718dd9