Set paths on PBXGroup types
Summary:
We're not setting this for now, so if you right click in a folder and say "Add File", it picks the project path as the root for putting the file which is not really ideal. Especially given that file structures in a project can be pretty deep, this means that adding files can be a chore.

To fix this, we need to assign a path to the groups. The best way to do this while maintaining backwards compatibility with projectv1 is to:

- Create a param on PBXProject to basically create the root path; this needs to be a relative path, so we can construct this using the project directory and then getting the relative path to the cell from it
- If this is set, for all subchildren, we can easily just pass the group name as the path component and set its SourceTree to GROUP to tell the project file that its folder reference should be relative to its parent group.

shipit-source-id: 74f5cb43f3198aaef3e16cfcd3c5972dd674563c