Remove use of `findFirst` in `BuildTargetSpec`
Summary:
The `filter` function of `BuildTargetSpec` will go through all the target nodes and look for one with the same unconfigured build target. It will then pick one arbitrarily and return it.

This doesn't seem right. On the one hand, `filter` lets you return multiple targets so there's nothing really stopping us from doing it, and it's clearly wrong to pick a target arbitrarily. On the other hand, `BuildTargetSpec` lists itself as `TargetType.SINGLE_TARGET`, so I can see an argument for why we would want to return only one (even if the process through which we choose that one is... questionable).

I'll be honest and say I don't fully understand the implications of this change, but I saw it and it looked wrong to me so I want to open it up for others to comment.

Reviewed By: stepancheg

shipit-source-id: 96c2ee7724c18c3f9a14b7efb69f35f6c0e6d0d7